feat: copy failed manifest from binding
