+++ pkg/controllers/workgenerator/controller.go
@@ -52,6 +52,9 @@ import (
 )
 
 var (
+	// maxFailedResourcePlacementLimit indicates the max number of failed resource placements to include in the status.
+	maxFailedResourcePlacementLimit = 100
+
 	errResourceSnapshotNotFound = errors.New("the master resource snapshot is not found")
 )
 
@@ -619,18 +622,18 @@ func setBindingStatus(works map[string]*fleetv1beta1.Work, resourceBinding *flee
 	resourceBinding.Status.FailedPlacements = nil
 	// collect and set the failed resource placements to the binding if not all the works are available
 	if appliedCond.Status != metav1.ConditionTrue || availableCond.Status != metav1.ConditionTrue {
-		failedResourcePlacements := make([]fleetv1beta1.FailedResourcePlacement, 0, controller.MaxFailedResourcePlacementLimit) // preallocate the memory
-		for _, work := range works {
-			if work.DeletionTimestamp != nil {
-				klog.V(2).InfoS("Ignoring the deleting work", "clusterResourceBinding", bindingRef, "work", klog.KObj(work))
+		failedResourcePlacements := make([]fleetv1beta1.FailedResourcePlacement, 0, maxFailedResourcePlacementLimit) // preallocate the memory
+		for _, w := range works {
+			if w.DeletionTimestamp != nil {
+				klog.V(2).InfoS("Ignoring the deleting work", "clusterResourceBinding", bindingRef, "work", klog.KObj(w))
 				continue // ignore the deleting work
 			}
-			failedManifests := controller.ExtractFailedResourcePlacementsFromWork(work)
+			failedManifests := extractFailedResourcePlacementsFromWork(w)
 			failedResourcePlacements = append(failedResourcePlacements, failedManifests...)
 		}
 		// cut the list to keep only the max limit
-		if len(failedResourcePlacements) > controller.MaxFailedResourcePlacementLimit {
-			failedResourcePlacements = failedResourcePlacements[0:controller.MaxFailedResourcePlacementLimit]
+		if len(failedResourcePlacements) > maxFailedResourcePlacementLimit {
+			failedResourcePlacements = failedResourcePlacements[0:maxFailedResourcePlacementLimit]
 		}
 		resourceBinding.Status.FailedPlacements = failedResourcePlacements
 		if len(failedResourcePlacements) > 0 {
@@ -642,10 +645,10 @@ func setBindingStatus(works map[string]*fleetv1beta1.Work, resourceBinding *flee
 func buildAllWorkAppliedCondition(works map[string]*fleetv1beta1.Work, binding *fleetv1beta1.ClusterResourceBinding) metav1.Condition {
 	allApplied := true
 	var notAppliedWork string
-	for _, work := range works {
-		if !condition.IsConditionStatusTrue(meta.FindStatusCondition(work.Status.Conditions, fleetv1beta1.WorkConditionTypeApplied), work.GetGeneration()) {
+	for _, w := range works {
+		if !condition.IsConditionStatusTrue(meta.FindStatusCondition(w.Status.Conditions, fleetv1beta1.WorkConditionTypeApplied), w.GetGeneration()) {
 			allApplied = false
-			notAppliedWork = work.Name
+			notAppliedWork = w.Name
 			break
 		}
 	}
@@ -734,6 +737,87 @@ func extractResFromConfigMap(uConfigMap *unstructured.Unstructured) ([]fleetv1be
 		return strings.Compare(string(obj1), string(obj2)) > 0
 	})
 	return manifests, nil
+}
+
+// extractFailedResourcePlacementsFromWork extracts the failed resource placements from the work.
+func extractFailedResourcePlacementsFromWork(work *fleetv1beta1.Work) []fleetv1beta1.FailedResourcePlacement {
+	appliedCond := meta.FindStatusCondition(work.Status.Conditions, fleetv1beta1.WorkConditionTypeApplied)
+	availableCond := meta.FindStatusCondition(work.Status.Conditions, fleetv1beta1.WorkConditionTypeAvailable)
+
+	// The applied condition and available condition are always updated in one call.
+	// It means the observedGeneration of these two are always the same.
+	// If IsConditionStatusFalse is true, means both are observing the latest work.
+	if !condition.IsConditionStatusFalse(appliedCond, work.Generation) &&
+		!condition.IsConditionStatusFalse(availableCond, work.Generation) {
+		return nil
+	}
+
+	// check if the work is generated by an enveloped object
+	envelopeType, isEnveloped := work.GetLabels()[fleetv1beta1.EnvelopeTypeLabel]
+	var envelopObjName, envelopObjNamespace string
+	if isEnveloped {
+		// If the work  generated by an enveloped object, it must contain those labels.
+		envelopObjName = work.GetLabels()[fleetv1beta1.EnvelopeNameLabel]
+		envelopObjNamespace = work.GetLabels()[fleetv1beta1.EnvelopeNamespaceLabel]
+	}
+	res := make([]fleetv1beta1.FailedResourcePlacement, 0, len(work.Status.ManifestConditions))
+	for _, manifestCondition := range work.Status.ManifestConditions {
+		failedManifest := fleetv1beta1.FailedResourcePlacement{
+			ResourceIdentifier: fleetv1beta1.ResourceIdentifier{
+				Group:     manifestCondition.Identifier.Group,
+				Version:   manifestCondition.Identifier.Version,
+				Kind:      manifestCondition.Identifier.Kind,
+				Name:      manifestCondition.Identifier.Name,
+				Namespace: manifestCondition.Identifier.Namespace,
+			},
+		}
+		if isEnveloped {
+			failedManifest.ResourceIdentifier.Envelope = &fleetv1beta1.EnvelopeIdentifier{
+				Name:      envelopObjName,
+				Namespace: envelopObjNamespace,
+				Type:      fleetv1beta1.EnvelopeType(envelopeType),
+			}
+		}
+
+		appliedCond = meta.FindStatusCondition(manifestCondition.Conditions, fleetv1beta1.WorkConditionTypeApplied)
+		// collect if there is an explicit fail
+		// The observedGeneration of the manifest condition is the generation of the applied manifest.
+		// The overall applied and available conditions are observing the latest work generation.
+		// So that the manifest condition should be latest, assuming they're populated by the work agent in one update call.
+		if appliedCond != nil && appliedCond.Status == metav1.ConditionFalse {
+			if isEnveloped {
+				klog.V(2).InfoS("Find a failed to apply enveloped manifest",
+					"manifestName", manifestCondition.Identifier.Name,
+					"group", manifestCondition.Identifier.Group,
+					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind,
+					"envelopeType", envelopeType, "envelopObjName", envelopObjName, "envelopObjNamespace", envelopObjNamespace)
+			} else {
+				klog.V(2).InfoS("Find a failed to apply manifest",
+					"manifestName", manifestCondition.Identifier.Name, "group", manifestCondition.Identifier.Group,
+					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind)
+			}
+			failedManifest.Condition = *appliedCond
+			res = append(res, failedManifest)
+			break
+		}
+		availableCond = meta.FindStatusCondition(manifestCondition.Conditions, fleetv1beta1.WorkConditionTypeAvailable)
+		if availableCond != nil && availableCond.Status == metav1.ConditionFalse {
+			if isEnveloped {
+				klog.V(2).InfoS("Find an unavailable enveloped manifest",
+					"manifestName", manifestCondition.Identifier.Name,
+					"group", manifestCondition.Identifier.Group,
+					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind,
+					"envelopeType", envelopeType, "envelopObjName", envelopObjName, "envelopObjNamespace", envelopObjNamespace)
+			} else {
+				klog.V(2).InfoS("Find an unavailable enveloped manifest",
+					"manifestName", manifestCondition.Identifier.Name, "group", manifestCondition.Identifier.Group,
+					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind)
+			}
+			failedManifest.Condition = *availableCond
+			res = append(res, failedManifest)
+		}
+	}
+	return res
 }
 
 // SetupWithManager sets up the controller with the Manager.
