+++ pkg/utils/controller/controller.go
@@ -14,8 +14,6 @@ import (
 	"time"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog/v2"
@@ -23,7 +21,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	fleetv1beta1 "go.goms.io/fleet/apis/placement/v1beta1"
-	"go.goms.io/fleet/pkg/utils/condition"
 	"go.goms.io/fleet/pkg/utils/controller/metrics"
 	"go.goms.io/fleet/pkg/utils/keys"
 	"go.goms.io/fleet/pkg/utils/labels"
@@ -34,11 +31,6 @@ const (
 	labelRequeueAfter = "requeue_after"
 	labelRequeue      = "requeue"
 	labelSuccess      = "success"
-)
-
-var (
-	// MaxFailedResourcePlacementLimit indicates the max number of failed resource placements to include in the status.
-	MaxFailedResourcePlacementLimit = 100
 )
 
 var (
@@ -351,85 +343,4 @@ func FetchAllClusterResourceSnapshots(ctx context.Context, k8Client client.Clien
 		return nil, NewExpectedBehaviorError(misMatchErr)
 	}
 	return resourceSnapshots, nil
-}
-
-// ExtractFailedResourcePlacementsFromWork extracts the failed resource placements from the work.
-func ExtractFailedResourcePlacementsFromWork(work *fleetv1beta1.Work) []fleetv1beta1.FailedResourcePlacement {
-	appliedCond := meta.FindStatusCondition(work.Status.Conditions, fleetv1beta1.WorkConditionTypeApplied)
-	availableCond := meta.FindStatusCondition(work.Status.Conditions, fleetv1beta1.WorkConditionTypeAvailable)
-
-	// The applied condition and available condition are always updated in one call.
-	// It means the observedGeneration of these two are always the same.
-	// If IsConditionStatusFalse is true, means both are observing the latest work.
-	if !condition.IsConditionStatusFalse(appliedCond, work.Generation) &&
-		!condition.IsConditionStatusFalse(availableCond, work.Generation) {
-		return nil
-	}
-
-	// check if the work is generated by an enveloped object
-	envelopeType, isEnveloped := work.GetLabels()[fleetv1beta1.EnvelopeTypeLabel]
-	var envelopObjName, envelopObjNamespace string
-	if isEnveloped {
-		// If the work  generated by an enveloped object, it must contain those labels.
-		envelopObjName = work.GetLabels()[fleetv1beta1.EnvelopeNameLabel]
-		envelopObjNamespace = work.GetLabels()[fleetv1beta1.EnvelopeNamespaceLabel]
-	}
-	res := make([]fleetv1beta1.FailedResourcePlacement, 0, len(work.Status.ManifestConditions))
-	for _, manifestCondition := range work.Status.ManifestConditions {
-		failedManifest := fleetv1beta1.FailedResourcePlacement{
-			ResourceIdentifier: fleetv1beta1.ResourceIdentifier{
-				Group:     manifestCondition.Identifier.Group,
-				Version:   manifestCondition.Identifier.Version,
-				Kind:      manifestCondition.Identifier.Kind,
-				Name:      manifestCondition.Identifier.Name,
-				Namespace: manifestCondition.Identifier.Namespace,
-			},
-		}
-		if isEnveloped {
-			failedManifest.ResourceIdentifier.Envelope = &fleetv1beta1.EnvelopeIdentifier{
-				Name:      envelopObjName,
-				Namespace: envelopObjNamespace,
-				Type:      fleetv1beta1.EnvelopeType(envelopeType),
-			}
-		}
-
-		appliedCond = meta.FindStatusCondition(manifestCondition.Conditions, fleetv1beta1.WorkConditionTypeApplied)
-		// collect if there is an explicit fail
-		// The observedGeneration of the manifest condition is the generation of the applied manifest.
-		// The overall applied and available conditions are observing the latest work generation.
-		// So that the manifest condition should be latest, assuming they're populated by the work agent in one update call.
-		if appliedCond != nil && appliedCond.Status == metav1.ConditionFalse {
-			if isEnveloped {
-				klog.V(2).InfoS("Find a failed to apply enveloped manifest",
-					"manifestName", manifestCondition.Identifier.Name,
-					"group", manifestCondition.Identifier.Group,
-					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind,
-					"envelopeType", envelopeType, "envelopObjName", envelopObjName, "envelopObjNamespace", envelopObjNamespace)
-			} else {
-				klog.V(2).InfoS("Find a failed to apply manifest",
-					"manifestName", manifestCondition.Identifier.Name, "group", manifestCondition.Identifier.Group,
-					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind)
-			}
-			failedManifest.Condition = *appliedCond
-			res = append(res, failedManifest)
-			break
-		}
-		availableCond = meta.FindStatusCondition(manifestCondition.Conditions, fleetv1beta1.WorkConditionTypeAvailable)
-		if availableCond != nil && availableCond.Status == metav1.ConditionFalse {
-			if isEnveloped {
-				klog.V(2).InfoS("Find an unavailable enveloped manifest",
-					"manifestName", manifestCondition.Identifier.Name,
-					"group", manifestCondition.Identifier.Group,
-					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind,
-					"envelopeType", envelopeType, "envelopObjName", envelopObjName, "envelopObjNamespace", envelopObjNamespace)
-			} else {
-				klog.V(2).InfoS("Find an unavailable enveloped manifest",
-					"manifestName", manifestCondition.Identifier.Name, "group", manifestCondition.Identifier.Group,
-					"version", manifestCondition.Identifier.Version, "kind", manifestCondition.Identifier.Kind)
-			}
-			failedManifest.Condition = *availableCond
-			res = append(res, failedManifest)
-		}
-	}
-	return res
 }
